complete various db controllers

+++ server/controllers/mousetrackingController.js
@@ -8,43 +8,36 @@ var model = require('../db/model');
 // }, { timestamps: false });
 
 // input should be of the following format:
-// { movement: 'abc', clicks: 'abc', urlchange: 'abc', test_id: 123 }
+// { movement: 'abc', clicks: 'abc', urlchange: 'abc', userId: 123, imageId: 123 }
 // output shall be of the following format:
 // { id: 123, movement: 'abc', clicks: 'abc', urlchange: 'abc' }
 var createMouseTracking = function (mouseTracking) {
   var params = {
+    userId: mouseTracking.userId,
+    imageId: mouseTracking.imageId,
     movement: mouseTracking.movement,
     clicks: mouseTracking.clicks,
     urlchange: mouseTracking.urlchange
   };
 
-  return model.sequelize.transaction(function (t) {
-    return model.MouseTracking.create(params, { transaction: t })
-      .then(function (newMouseTracking) {
-        var params = { test_id: mouseTracking.test_id, mousetracking_id: newMouseTracking.get('id') };
-
-        return model.MouseTrackingTest.create(params, { transaction: t })
-          .then(function (mouseTrackingTest) {
-            if (mouseTrackingTest === null) {
-              throw (new Error ('Error! Unable to create mousetracking_test join!'));
-            } else {
-              return newMouseTracking;
-            }
-          });
-      });
-  });
+  return model.MouseTracking.create(params)
+    .then(function (newMouseTracking) {
+      var params = { test_id: mouseTracking.test_id, mousetracking_id: newMouseTracking.get('id') };
+    });
 };
 
 // input should be of the following format:
-// { id: 123 }
+// { userId: 123, imageId: 123 }
 // output shall be of the following format:
 // { id: 123, movement: 'abc', clicks: 'abc', urlchange: 'abc' }
+
 var retrieveMouseTracking = function (mouseTracking) {
+  console.log('mouse tracking input:', mouseTracking)
   return model.MouseTracking.findAll({
-    where: mouseTracking.image,
+    where: { imageId: mouseTracking.image },
     include: [{
       model: model.User,
-      where: mouseTracking.user,
+      where: { id: mouseTracking.user },
       attributes: [ 'id', 'email' ],
       include: [{
         model: model.Project,
@@ -52,7 +45,7 @@ var retrieveMouseTracking = function (mouseTracking) {
           model: model.Test,
           include: [{
             model: model.Image,
-            where: { id: mouseTracking.image.imageId }
+            where: { id: mouseTracking.image }
           }]
         }]
       }]
@@ -74,14 +67,38 @@ var retrieveMouseTracking = function (mouseTracking) {
 var updateMouseTracking = function (mouseTracking) {
   var params = { id: mouseTracking.id };
 
-  return model.MouseTracking.update(mouseTracking, {
-    where: params
+  return model.MouseTracking.findAll({
+    where: { id: mouseTracking.image },
+    include: [{
+      model: model.User,
+      where: { id: mouseTracking.user },
+      attributes: [ 'id', 'email' ],
+      include: [{
+        model: model.Project,
+        include: [{
+          model: model.Test,
+          include: [{
+            model: model.Image,
+            where: { id: mouseTracking.image }
+          }]
+        }]
+      }]
+    }]
   })
-  .spread(function (updated) {
-    if (updated === 0) {
-      throw (new Error ('Error! Comment update failed!'));
+  .spread(function (result) {
+    if (result.user.projects[0].projectUser.get('role') === 'owner') {
+      return model.MouseTracking.update(mouseTracking, {
+        where: params
+      })
+      .spread(function (updated) {
+        if (updated === 0) {
+          throw (new Error ('Error! Comment update failed!'));
+        } else {
+          return comment;
+        }
+      });
     } else {
-      return comment;
+      throw (new Error ('Error! Insufficient permissions to modify this entry!'))
     }
   });
 };
@@ -91,14 +108,38 @@ var updateMouseTracking = function (mouseTracking) {
 // output shall be of the following format:
 // 1
 var deleteMouseTracking = function (mouseTracking) {
-  return model.MouseTracking.destroy({
-    where: mouseTracking
+  return model.MouseTracking.findAll({
+    where: { id: mouseTracking.image },
+    include: [{
+      model: model.User,
+      where: { id: mouseTracking.user },
+      attributes: [ 'id', 'email' ],
+      include: [{
+        model: model.Project,
+        include: [{
+          model: model.Test,
+          include: [{
+            model: model.Image,
+            where: { id: mouseTracking.image }
+          }]
+        }]
+      }]
+    }]
   })
-  .then(function (deleted) {
-    if (deleted === 0) {
-      throw (new Error ('Error! Mouse tracking delete failed!'));
+  .spread(function (result) {
+    if (result.user.projects[0].projectUser.get('role') === 'owner') {
+      return model.MouseTracking.destroy({
+        where: mouseTracking
+      })
+      .then(function (deleted) {
+        if (deleted === 0) {
+          throw (new Error ('Error! Mouse tracking delete failed!'));
+        } else {
+          return deleted;
+        }
+      });
     } else {
-      return deleted;
+      throw (new Error ('Error! Insufficient permissions to modify this entry!'))
     }
   });
 };
